[APP][GUARDIAN] Added fire detection test window                 added sensibility editing in movement test window                 fix some crash

+++ Assets/Apps/Guardian/Scripts/Behaviour/MenuStates/DebugThermicState.cs
@@ -44,7 +44,7 @@ namespace BuddyApp.Guardian
             }
             else if (mHasOpenedWindow) 
             {
-                int[] lThermicMatrix = new int[64];
+                float[] lThermicMatrix = new float[64];
                 OpenCVUnity.Mat lMatConverted = new OpenCVUnity.Mat(8,8, OpenCVUnity.CvType.CV_32SC1);
                 Buddy.Sensors.ThermalCamera.Frame.put(0, 0, lThermicMatrix); 
                 //.convertTo(lMatConverted, OpenCVUnity.CvType.CV_32SC1);
